// planning-idsdp-pps-frontend/src/app/main/modules/project-management/models/pm-view-schedule.model.ts
export class PmViewSelectionModel {

    id: number;
    dppTappMasterId: number;
    dppTappMasterUuid: string;
    uuid: string;
    taskDetails: string;
    location: string;
    startDate: Date;
    startTime: string;
    endDate: Date;
    endTime: string;
    remarks: string;
    attachedFile: File;
    attachmentName: string;
    attachment: any;

}